Create Question 3
Yes, by default, Django signals run in the same database transaction as the caller. This means that if the transaction is rolled back, the operations performed in the signal handler will also be rolled back.
In this example, the signal handler tries to update the instance, but since we force a rollback, the change will not persist, demonstrating that signals run in the same transaction.